requirements.txt
python-dotenv
langchain_groq
docx